+++ policy/nomad/validate_test.go
@@ -5,7 +5,7 @@ import (
 	"fmt"
 	"testing"
 
-	"github.com/hashicorp/nomad-autoscaler/helper/ptr"
+	"github.com/hashicorp/nomad-autoscaler/sdk/helper/ptr"
 	"github.com/hashicorp/nomad/api"
 	"github.com/stretchr/testify/assert"
 )
