+++ policyeval/broker.go
@@ -8,8 +8,8 @@ import (
 	"time"
 
 	hclog "github.com/hashicorp/go-hclog"
-	"github.com/hashicorp/nomad-autoscaler/helper/uuid"
 	"github.com/hashicorp/nomad-autoscaler/sdk"
+	"github.com/hashicorp/nomad-autoscaler/sdk/helper/uuid"
 )
 
 // Broker stores, dedups and control access to policy evaluation requests.
