+++ policyeval/broker_test.go
@@ -6,8 +6,8 @@ import (
 	"time"
 
 	hclog "github.com/hashicorp/go-hclog"
-	"github.com/hashicorp/nomad-autoscaler/helper/uuid"
 	"github.com/hashicorp/nomad-autoscaler/sdk"
+	"github.com/hashicorp/nomad-autoscaler/sdk/helper/uuid"
 	"github.com/stretchr/testify/assert"
 )
