+++ sdk/helper/scaleutils/filter.go
