+++ sdk/helper/scaleutils/request.go
